Bugfix: surface heat capacity
The variable land_sea_heat_capacity was initialized with the wrong size.
This lead to processors not having the right value, depending on the number of CPUs.

Also, as a minor change, re-organized the 'if' tree for the different surface
heat capacity options. Removed option 'mjtrop', because this would mean that it's
impossible to have separate tropical and land heat capacities.
Check is now first where the option is 'input', and if not, goes through all the
options of different tropical heat capacity, and then either 'zsurf', or 'lonlat'.