Adding back mppnccombine.x
The executable mppnccombine.x probably shouldn't be here, as in theory it could be compiled from the source. However, I can't get it to compile on gv3 or gv2, and as a temporary measure (until we're all using the python runscripts, when mppnccombine.x isn't required?) I thought I could add it back.