Compilation on Mac
Cores on a Mac cannot be tied in the same way that they can on linux, so core affinity is not possible.  Here I am just stubbing out the core affinity functions to mimic the linux case where it cannot get the affinity info.

More info in links inline in `affinity.c`.