Bug fix from Paul O'Gorman, recieved by Geoff on August 3rd 2018.
Paul's explanation is below:
******
I found a problem with the old boundary-layer diffusivity module that
is used in the  FMS "gray-radiation" code.

The issue is that the diffusivity can be set to non-zero spurious
values above the boundary layer when the boundary-layer height varies
a lot over a processor domain. Fortunately this doesn't seem to
strongly affect the climate in the kinds of simulations I have run
with the GCM. But a knock-on effect is that simulation output can
depend on the processor decomposition used.
******